+++ tests/unit/test_search.py
@@ -5,6 +5,7 @@ from searx.preferences import Preferences
 from searx.engines import engines
 
 import searx.search
+from searx.search import SearchQuery
 
 
 SAFESEARCH = 0
@@ -40,53 +41,53 @@ class SearchTestCase(SearxTestCase):
 
     def test_timeout_simple(self):
         searx.search.max_request_timeout = None
-        search_query = searx.query.SearchQuery('test', [{'category': 'general', 'name': PUBLIC_ENGINE_NAME}],
-                                               ['general'], 'en-US', SAFESEARCH, PAGENO, None, None,
-                                               preferences=Preferences(['oscar'], ['general'], engines, []))
+        search_query = SearchQuery('test', [{'category': 'general', 'name': PUBLIC_ENGINE_NAME}],
+                                   ['general'], 'en-US', SAFESEARCH, PAGENO, None, None,
+                                   preferences=Preferences(['oscar'], ['general'], engines, []))
         search = searx.search.Search(search_query)
         search.search()
         self.assertEqual(search.actual_timeout, 3.0)
 
     def test_timeout_query_above_default_nomax(self):
         searx.search.max_request_timeout = None
-        search_query = searx.query.SearchQuery('test', [{'category': 'general', 'name': PUBLIC_ENGINE_NAME}],
-                                               ['general'], 'en-US', SAFESEARCH, PAGENO, None, 5.0,
-                                               preferences=Preferences(['oscar'], ['general'], engines, []))
+        search_query = SearchQuery('test', [{'category': 'general', 'name': PUBLIC_ENGINE_NAME}],
+                                   ['general'], 'en-US', SAFESEARCH, PAGENO, None, 5.0,
+                                   preferences=Preferences(['oscar'], ['general'], engines, []))
         search = searx.search.Search(search_query)
         search.search()
         self.assertEqual(search.actual_timeout, 3.0)
 
     def test_timeout_query_below_default_nomax(self):
         searx.search.max_request_timeout = None
-        search_query = searx.query.SearchQuery('test', [{'category': 'general', 'name': PUBLIC_ENGINE_NAME}],
-                                               ['general'], 'en-US', SAFESEARCH, PAGENO, None, 1.0,
-                                               preferences=Preferences(['oscar'], ['general'], engines, []))
+        search_query = SearchQuery('test', [{'category': 'general', 'name': PUBLIC_ENGINE_NAME}],
+                                   ['general'], 'en-US', SAFESEARCH, PAGENO, None, 1.0,
+                                   preferences=Preferences(['oscar'], ['general'], engines, []))
         search = searx.search.Search(search_query)
         search.search()
         self.assertEqual(search.actual_timeout, 1.0)
 
     def test_timeout_query_below_max(self):
         searx.search.max_request_timeout = 10.0
-        search_query = searx.query.SearchQuery('test', [{'category': 'general', 'name': PUBLIC_ENGINE_NAME}],
-                                               ['general'], 'en-US', SAFESEARCH, PAGENO, None, 5.0,
-                                               preferences=Preferences(['oscar'], ['general'], engines, []))
+        search_query = SearchQuery('test', [{'category': 'general', 'name': PUBLIC_ENGINE_NAME}],
+                                   ['general'], 'en-US', SAFESEARCH, PAGENO, None, 5.0,
+                                   preferences=Preferences(['oscar'], ['general'], engines, []))
         search = searx.search.Search(search_query)
         search.search()
         self.assertEqual(search.actual_timeout, 5.0)
 
     def test_timeout_query_above_max(self):
         searx.search.max_request_timeout = 10.0
-        search_query = searx.query.SearchQuery('test', [{'category': 'general', 'name': PUBLIC_ENGINE_NAME}],
-                                               ['general'], 'en-US', SAFESEARCH, PAGENO, None, 15.0,
-                                               preferences=Preferences(['oscar'], ['general'], engines, []))
+        search_query = SearchQuery('test', [{'category': 'general', 'name': PUBLIC_ENGINE_NAME}],
+                                   ['general'], 'en-US', SAFESEARCH, PAGENO, None, 15.0,
+                                   preferences=Preferences(['oscar'], ['general'], engines, []))
         search = searx.search.Search(search_query)
         search.search()
         self.assertEqual(search.actual_timeout, 10.0)
 
     def test_query_private_engine_without_token(self):
-        search_query = searx.query.SearchQuery('test', [{'category': 'general', 'name': PRIVATE_ENGINE_NAME}],
-                                               ['general'], 'en-US', SAFESEARCH, PAGENO, None, 2.0,
-                                               preferences=Preferences(['oscar'], ['general'], engines, []))
+        search_query = SearchQuery('test', [{'category': 'general', 'name': PRIVATE_ENGINE_NAME}],
+                                   ['general'], 'en-US', SAFESEARCH, PAGENO, None, 2.0,
+                                   preferences=Preferences(['oscar'], ['general'], engines, []))
         search = searx.search.Search(search_query)
         results = search.search()
         self.assertEqual(results.results_length(), 0)
@@ -94,9 +95,9 @@ class SearchTestCase(SearxTestCase):
     def test_query_private_engine_with_incorrect_token(self):
         preferences_with_tokens = Preferences(['oscar'], ['general'], engines, [])
         preferences_with_tokens.parse_dict({'tokens': 'bad-token'})
-        search_query = searx.query.SearchQuery('test', [{'category': 'general', 'name': PRIVATE_ENGINE_NAME}],
-                                               ['general'], 'en-US', SAFESEARCH, PAGENO, None, 2.0,
-                                               preferences=preferences_with_tokens)
+        search_query = SearchQuery('test', [{'category': 'general', 'name': PRIVATE_ENGINE_NAME}],
+                                   ['general'], 'en-US', SAFESEARCH, PAGENO, None, 2.0,
+                                   preferences=preferences_with_tokens)
         search = searx.search.Search(search_query)
         results = search.search()
         self.assertEqual(results.results_length(), 0)
@@ -104,28 +105,28 @@ class SearchTestCase(SearxTestCase):
     def test_query_private_engine_with_correct_token(self):
         preferences_with_tokens = Preferences(['oscar'], ['general'], engines, [])
         preferences_with_tokens.parse_dict({'tokens': 'my-token'})
-        search_query = searx.query.SearchQuery('test', [{'category': 'general', 'name': PRIVATE_ENGINE_NAME}],
-                                               ['general'], 'en-US', SAFESEARCH, PAGENO, None, 2.0,
-                                               preferences=preferences_with_tokens)
+        search_query = SearchQuery('test', [{'category': 'general', 'name': PRIVATE_ENGINE_NAME}],
+                                   ['general'], 'en-US', SAFESEARCH, PAGENO, None, 2.0,
+                                   preferences=preferences_with_tokens)
         search = searx.search.Search(search_query)
         results = search.search()
         self.assertEqual(results.results_length(), 1)
 
     def test_external_bang(self):
-        search_query = searx.query.SearchQuery('yes yes',
-                                               [{'category': 'general', 'name': PUBLIC_ENGINE_NAME}],
-                                               ['general'], 'en-US', SAFESEARCH, PAGENO, None, None,
-                                               preferences=Preferences(['oscar'], ['general'], engines, [],),
-                                               external_bang="yt")
+        search_query = SearchQuery('yes yes',
+                                   [{'category': 'general', 'name': PUBLIC_ENGINE_NAME}],
+                                   ['general'], 'en-US', SAFESEARCH, PAGENO, None, None,
+                                   preferences=Preferences(['oscar'], ['general'], engines, [],),
+                                   external_bang="yt")
         search = searx.search.Search(search_query)
         results = search.search()
         # For checking if the user redirected with the youtube external bang
         self.assertTrue(results.redirect_url is not None)
 
-        search_query = searx.query.SearchQuery('youtube never gonna give you up',
-                                               [{'category': 'general', 'name': PUBLIC_ENGINE_NAME}],
-                                               ['general'], 'en-US', SAFESEARCH, PAGENO, None, None,
-                                               preferences=Preferences(['oscar'], ['general'], engines, []),)
+        search_query = SearchQuery('youtube never gonna give you up',
+                                   [{'category': 'general', 'name': PUBLIC_ENGINE_NAME}],
+                                   ['general'], 'en-US', SAFESEARCH, PAGENO, None, None,
+                                   preferences=Preferences(['oscar'], ['general'], engines, []),)
 
         search = searx.search.Search(search_query)
         results = search.search()
